% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{add_ccf_to_output}
\alias{add_ccf_to_output}
\title{Function to motify the output of SVClone to include the CCF into text files}
\usage{
add_ccf_to_output(base_dir, samples, test_run = T)
}
\arguments{
\item{base_dir}{The directory where the subfolders are kept}

\item{samples}{A vector of identifiers for the samples which are used in the 
folder structure of SVClone}

\item{test_run}{A boolean value which, if True, prevents the function from 
writing any files.}
}
\description{
This function reads in files that are stored as .RData (_ccube_sv_results.RData)
and pulls out the cancer cell fraction (CCF) values and then appends them into
_subclonal_structure.txt files, then writes a new file with this information
at _subclonal_structure_with_CCF_alan.txt files.
}
