% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{call_svs_per_sample}
\alias{call_svs_per_sample}
\title{Goes through all SVs and assigns names to SV}
\usage{
call_svs_per_sample(sample_list, threshold)
}
\arguments{
\item{sample_list}{A list where the indices of the list correspond to the
sample. This can be generated using the make_sample_list function}

\item{threshold}{The threshold for which SVs should be called the same}
}
\value{
A list including all_svs_per_sample, master_list, and sample_list
}
\description{
If two SVs, x and y, in the sample_list have pos1 and pos2 within threshold of each
other, then the two SVs are given the same name. If a third SV, z, is within
threshold of y, but not x, it is still given the same name. Names are 
assigned sequentially based on the order they appear in sample_list
}
