% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{clonality_plot}
\alias{clonality_plot}
\title{Function to create a clonality plot}
\usage{
clonality_plot(matched_sv_fns)
}
\arguments{
\item{matched_sv_fns}{A vector of file names for the matched output.}
}
\description{
Creates a plot where each point is a structural variant, and the x-axis is
the number of samples for which the SV was called subclonal, and the y-axis
is the number of samples for which the SV was called clonal
}
