% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{make_merged_table}
\alias{make_merged_table}
\title{A function to create a table with all of the SVs within one patient}
\usage{
make_merged_table(sample_list, threshold = 500, CCF_cutoff = 0.9)
}
\arguments{
\item{sample_list}{A sample_list from the make_sample_list function.}

\item{threshold}{The base pair threshold for calling SVs the same}

\item{CCF_cutoff}{The cutoff for CCF to consider an SV as clonal}
}
\description{
Takes a sample list, calls the call_svs_per_sample function, and then 
calls each SV as either clonal if it has cancer cell fraction (CCF) > 0.9 
(by default) or subclonal otherwise on a sample level.
On a patient-level, the SVs are called as the following.
Founder clone: clonal in all samples
Private clonal: clonal in exactly 1 sample, subclonal in 0.
Shared subclonal: clonal in no samples, subclonal in more than 1.
Private subclonal: clonal in no samples, subclonal in exactly 1 sample.
Shared clonal/subclonal: otherwise
}
