% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{make_sample_list}
\alias{make_sample_list}
\title{Finds all the samples in the folder with sample_name and reads them into a sample_list format}
\usage{
make_sample_list(
  sample_ids,
  cluster_certainty_fns,
  cluster_ccf_fns,
  sv_info_fns
)
}
\arguments{
\item{sample_ids}{A vector of the identifying names for each of the samples}

\item{cluster_certainty_fns}{A vector of the file names (fns) for the cluster
certainty files which are output by SVClone. For example
(IDENTIFIER)_cluster_certainty.txt}

\item{cluster_ccf_fns}{A vector of the file names (fns) for the cancer cell
fraction (CCF) files from SVClone. For example
(IDENTIFIER)_subclonal_structure_with_CCF_alan.txt}

\item{sv_info_fns}{A vector of the file names (fns) for the additional info
files from SVClone. For example
(IDENTIFIER)__filtered_svs.tsv}
}
\value{
a list of samples that binds together the
cluster certainty file which is output from SVClone, and the additional information
then makes the index of the list the name of the sample.
}
\description{
The sample list format is a list of samples that binds together the
cluster certainty file which is output from SVClone, and the additional information
then makes the index of the list the name of the sample.
}
