% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{within_threshold}
\alias{within_threshold}
\title{Checks if the start1, end1, start2, end2 are within the threshold}
\usage{
within_threshold(start1, end1, start2, end2, threshold)
}
\arguments{
\item{start1}{position 1 for first sample}

\item{end1}{position 2 for first sample}

\item{start2}{position 1 for second sample}

\item{end2}{position 2 for second sample}

\item{threshold}{how far away is acceptable for start1 and start2}
}
\description{
This is just a macro for checking if
(abs(start1 - start2) <= threshold & abs(end1 - end2) <= threshold)
is true or not. If it's true, we return True.
}
