% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmerge.R
\name{write_merged_tables}
\alias{write_merged_tables}
\title{Takes SVClone output from different tissues in the same patient and merges them}
\usage{
write_merged_tables(
  sample_ids,
  patient_ids,
  cluster_certainty_fns,
  sv_info_fns,
  cluster_ccf_fns,
  merged_table_output_folder,
  threshold = 500
)
}
\description{
This function takes in the the outputs from SVClone and combines them. It 
requires that the function to generate the cancer cell fraction (CCF) and 
put it into the SVClone output files has already been run.
}
